+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Change Log
 ### Improvements
 
 - Change counting method of the cursor location and column to 1-based (formerly 0-based).
+- Stop showing syntax highlighting indicator dialog.
 - Optimize syntax highlight application so that the time displaying the rainbow cursor reduces significantly with a large document.
 - Move focus to the editor when lines are selected by clicking line numbers.
 - Apply update of “Show invisible characters” option in the preferences immediately to opened documents.